+++ lib/src/animate.dart
@@ -108,7 +108,7 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
   /// Creates an Animate instance that will manage a list of effects and apply
   /// them to the specified child.
   Animate({
-    Key? key,
+    super.key,
     this.child = const SizedBox.shrink(),
     List<Effect>? effects,
     this.onInit,
@@ -119,10 +119,9 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
     this.controller,
     this.adapter,
     this.target,
-    this.isAnimationEnabled = true,
+    this.enabled = true,
   })  : autoPlay = autoPlay ?? true,
-        delay = delay ?? Duration.zero,
-        super(key: key) {
+        delay = delay ?? Duration.zero {
     warn(
       autoPlay != false || onPlay == null,
       'Animate.onPlay is not called when Animate.autoPlay=false',
@@ -147,9 +146,9 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
   /// If you don't want to play the animation and just return the child
   /// for some reasons and you don't want to create helper widget or function
   /// or builder that help to achieve that without too duplication
-  /// then please pass false to [isAnimationEnabled] as a value to return the
+  /// then please pass false to [enabled] as a value to return the
   /// child directly
-  final bool isAnimationEnabled;
+  final bool enabled;
 
   /// Called immediately after the controller is fully initialized, before
   /// the [Animate.delay] or the animation starts playing (see: [onPlay]).
@@ -377,7 +376,7 @@ class _AnimateState extends State<Animate> with SingleTickerProviderStateMixin {
   Widget build(BuildContext context) {
     Widget child = widget.child, parent = child;
 
-    if (!widget.isAnimationEnabled) {
+    if (!widget.enabled) {
       return child;
     }
     ReparentChildBuilder? reparent = Animate.reparentTypes[child.runtimeType];
@@ -403,6 +402,7 @@ extension AnimateWidgetExtensions on Widget {
     AnimationController? controller,
     Adapter? adapter,
     double? target,
+    bool enabled = true,
   }) =>
       Animate(
         key: key,
@@ -415,6 +415,7 @@ extension AnimateWidgetExtensions on Widget {
         controller: controller,
         adapter: adapter,
         target: target,
+        enabled: enabled,
         child: this,
       );
 }